Create Containerfile for Edge TPU on Ubuntu
After trying to build libraries for the Coral USB Edge TPU on
Fedora, I gave up and resorted to trying to run it in an
Ubuntu container instead.  There was a lot of agony trying to get
the right environment, but I settled on some Wheel files built
by Github user oberluz for pycoral and tflite_runtime.

It appears that Google has stopped doing releases for pycoral,
and is letting the community run with it now.