Moving model training code to training directory.
Mostly a lot of cleanup to separate the training code from the
inferencing code.  The goal is to be able to deploy training and
inferencing completely separately.  This matches the expected deployment
pattern for the project where training will happen in one environment
and inferencing will happen in another environment.

It also made sense to move the config info to this training directory
since the configs were all about training and nothing about inferencing.

And, removing a couple pycache files that accidentally slipped into an
earlier commit.  BE GONE!