A lot of changes to the inferencing code
Since loading the libedgetpu interpreter takes so long, I changed the
inferencing logic to be able to perform inferencing all images in a
directory instead of a single image.  But, it also meant moving the
"handle a squirrel or no-squirrel image" in the python code instead of
externally.  I don't like this as much, because I really wanted the
inferencing to be as simple as possible.  But, whatever.